+++ csrc/include/natten/cuda/fmha_hopper/collective/fmha_collective_tma.hpp
@@ -74,7 +74,7 @@ struct FmhaMainloopTmaSm90 {
   using TileShapeQK = TileShape;
   using TileShapePV = decltype(select<0, 2, 1>(TileShapeQK{}));
 
-  using LayoutQKV = cute::tuple<int, _1, cute::tuple<int, int>>;
+  using LayoutQKV = cute::tuple<int, _1, cute::tuple<int64_t, int>>;
   using LayoutQ = LayoutQKV;
   using LayoutK = LayoutQKV;
   using LayoutV = LayoutQKV;
